Fix/improve gometalinter pre commit hook: fail on ineffassign
Ineffectual assignment is a bug! (I don't know of any case it cannot be
a bug for now, at the very least) And it is very time consuming when you
are presented with error, but that error is not an actual error (this
happens when you forget to handle `err`, see the code below). This drove
me nuts on my job so I'm trying to fail on pre commit whenever there's
single ineffectual assignment in a project.

Example:

```
func doStuff() error {
  foo, err := saveFoo()
  // oops, I forgot to handle err since I don't care about the code quality
  bar, err := getBar()
  if err != nil {
    return errors.Wrap(err, "failed to get bar")
  }
  doOtherStuff()
  return nil
}
```

Now imagine you are presented with the following log:
```
failed to get bar: dependency is not satisfied for bar
```

Imagine `getBar` depends on previous `saveFoo` completes successfully.
Imagine `saveFoo` has failed, and that led to `getBar` failing,
producing the above error message. This is not an actual error you want
to see. If `err` from `saveFoo` was handled you should exactly know the
cause is `saveFoo`.