surfingkeys: Disable Surfingkeys on Google Docs
Also, use <C-h> and <C-l> to move to next/previous tabs. This is because
<C-p> is intercepted by Google Docs, showing print dialog (sigh).
(Also I cannot avoid using Google Docs, unfortunately)

It helps to show tabs at bottom (via Vivaldi settings), which resembles
tmux windows. I move windows on tmux with <C-h>/<C-l>, so it's not too
hard to get used to it I believe.